docker: wait for interfaces to be ready
At entrypoint time, kernel network interfaces might not be ready for
the clsact qdisc to be attached. This manifests as:

```
Error: Exclusivity flag on, cannot modify.
```

The solution is to wait. This commit retries adding the tc qdisc
up to N_TIMES=5, and adds a delay of 5s between executions.