sfunnel.c: add rule engine
Prior to this commit, the type of traffic matched and the actions
were backed in the code.

This commit allows to define a set of rules match+actions, so that
the same codebase can be used to match a number of protocols and
perform several actions.

Rules are matched in **strict order**, which allows certain matching

Please note, sfunnel - as of today - is not designed to handle a lot
of rules (you would typically use it with 5 or less rules). The lookup
is simple and linear.

* Supported matches:
  * ip saddr/daddr: source/destination CIDR
  * tcp sport/dport: source/destination TCP Port
  * udp sport/dport: source/destination UDP Port
* Supported actions:
  * funnel: which allows to choose protocol {tcp,udp} and L4 sport/dport
    - NOTE: TCP flags cannot be set at this point
  * unfunnel: reverse operation
  * accept: do not touch the packet (carve-out)
  * drop: drop the packet
  * dnat: defined but NOT SUPPORTED

Next steps:

* Add more coverage to `gen.py` and possible generated rules.
* Enable the docker container to define rules both as a file (configmap)
  and via env variable.
* Support matching and setting fwmark (with mask).