test/cni|ci: add CNI performance tests
This commit adds a very simple performance testing on all CNIs that
evaluate the sfunnel example.

The tests is a simple `iperf` between a client and the sfunneled
service in K8s. This test runs as a last step of the CNI CI jobs,
and a Markdown report (to be improved) is generated.

This commit along with the next one have uncovered the issues with
GSO as reported in issue <put issue>.