test/cni: work-around conntrack lingering rules
If not enough time passes between the affinity and the perf test
(E.g. in the CI), the perf test fails on all CNIs (but Calico?) due
to lingering DNAT conntrack rules that map to old endpoints.

This seems a bug in either MetalLB or kubeproxy, as it points to
IP addresses that are not anymore valid.