src: ungso traffic before un/funneling
This commit workarounds the issues with GSO (`gso_type`) as described
in #11 and #12 by forcing traffic go through a new pair of veth interfaces,
`_seg`/`_seg_pair` with GSO disabled, _prior_ to be funneled/unfunneled.
bpf_redirect() is used to send packets from the ingress or egress interface
to the `_seg` interface.

This workarounds the issues at the expense of performance for the
funneled/unfunneled traffic (no GSO), but without affecting the rest of
the traffic.

Further commits will refine this approach, and will adjust MTU (PMTUD)
accordingly.

Thank you Daniel Borkman for the suggestion!

Fixes #11, #12