Dynamically adjust MTU using PMTUD when funnneling
This patch generated ICMP Fragmentation needed packets when funneling
rules would exceed the effective MTU towards a destination. This
ensures no MTU issues arise when funneling.

To do that, sfunnel uses an unpinned PMTUD map to store state on which
flows have been PMTUD adjusted or not. Note that this commit does NOT
intercept (yet) network originated PMTUD packets,

In addition, this commit bumps kernel requirements to 5.12 or above,
since it requires the `mtu_result` as part of the FIB lookup BPF helper.

Superseeds #3, implements #4